requirements.txt
fastapi
pydantic
google-genai
uvicorn
requests